+++ gd-gltg/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GLGVO-630] (https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-630)
 - Added better explanation of the Avg. Annual Nutrient yield graph
   [GLGVO-640](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-640)
+- Added overall nitrate load text for IL drainage
+  [GLGVO-641](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-641)
 
 ### Changed
 - Create Intervals as a Reusable Component